Server handles "get"-Request instead of "post" (again).
In the end, "get" is closer to the Restful-definition (retrieving a resource).
It works better with HTTP-caching mechanisms. Finally, "post" does not offer
a significant advantage over "get". Simple key-value objects are probably sufficient for
configuring the conversion. Such objects can easily be mapped onto query-strings, so
there is no need for posting a JSON object.